# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: >-
    Aircraft Engine MRO Decision Support API. Ingests engine condition data,
    reasons over it with a domain ontology, knowledge graph and rules engine,
    and produces traceable MRO Work Recommendations for human planners and for
    SAP S/4HANA Cloud (mock mode).
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rules/{id}:
    get:
      operationId: getRule
      tags: [rules]
      summary: Get one rule
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Edit a rule (thresholds, cycles, enable/disable)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Rule:
      type: object
      required:
        - id
        - name
        - description
        - failureMode
        - parameter
        - operator
        - threshold
        - consecutiveCycles
        - severity
        - enabled
        - autoApprove
        - component
        - regulatoryRefs
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        failureMode: { type: string }
        parameter: { type: string }
        operator:
          type: string
          enum: [lt, gt, trend_down, trend_up]
        threshold: { type: number }
        consecutiveCycles: { type: integer }
        severity: { type: number }
        enabled: { type: boolean }
        autoApprove: { type: boolean }
        component: { type: string }
        recommendedTaskCode: { type: ["string", "null"] }
        regulatoryRefs:
          type: array
          items: { type: string }
    RuleUpdate:
      type: object
      properties:
        threshold: { type: number }
        consecutiveCycles: { type: integer }
        severity: { type: number }
        enabled: { type: boolean }
        autoApprove: { type: boolean }
...